+++ src/gui/widgets/node_menu.cpp
@@ -11,6 +11,8 @@
 #include "model/shapes/repeater.hpp"
 #include "model/shapes/trim.hpp"
 #include "model/shapes/text.hpp"
+#include "model/shapes/inflate_deflate.hpp"
+#include "model/shapes/round_corners.hpp"
 
 #include "model/assets/assets.hpp"
 
@@ -217,6 +219,8 @@ void actions_group(QMenu* menu, GlaxnimateWindow* window, model::Group* group)
     menu_add->addSeparator();
     add_child_action<model::Trim>(menu_add, group);
     add_child_action<model::Repeater>(menu_add, group);
+    add_child_action<model::InflateDeflate>(menu_add, group);
+    add_child_action<model::RoundCorners>(menu_add, group);
 
     menu->addSeparator();